requirements.txt
mariadb
atlassian-python-api
